Address another fixme
